+++ Connect-Office365Services/Public/Get-OnPremisesCredentials.ps1
@@ -1,9 +1,9 @@
-function Get-OnPremisesCredentials {
-    $local:prevuser= if ($script:myOffice365Services['OnPremisesCredentials']) {
-        $script:myOffice365Services['OnPremisesCredentials'].UserName
-    } 
-    else { 
+function Get-OnPremisesCredential {
+    $local:prevuser= if ($script:myOffice365Services['OnPremisesCredential']) {
+        $script:myOffice365Services['OnPremisesCredential'].UserName
+    }
+    else {
         ''
     }
-    $script:myOffice365Services['OnPremisesCredentials'] = Get-Credential -UserName $local:prevUser -Message 'Please enter your on-premises credentials' -Title 'On-Premises Credentials' -ErrorAction SilentlyContinue
+    $script:myOffice365Services['OnPremisesCredential'] = Get-Credential -UserName $local:prevUser -Message 'Please enter your on-premises credentials' -Title 'On-Premises Credentials' -ErrorAction SilentlyContinue
 }